Split compilation of dense kernels
