+++ common/unified/matrix/dense_kernels.cpp
@@ -38,9 +38,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 #include "common/unified/base/kernel_launch.hpp"
-#include "common/unified/base/kernel_launch_reduction.hpp"
 #include "core/base/mixed_precision_types.hpp"
-#include "core/components/prefix_sum_kernels.hpp"
 
 
 namespace gko {
@@ -240,176 +238,6 @@ void sub_scaled_diag(std::shared_ptr<const DefaultExecutor> exec,
 }
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_SUB_SCALED_DIAG_KERNEL);
-
-
-template <typename ValueType>
-void compute_dot(std::shared_ptr<const DefaultExecutor> exec,
-                 const matrix::Dense<ValueType>* x,
-                 const matrix::Dense<ValueType>* y,
-                 matrix::Dense<ValueType>* result, array<char>& tmp)
-{
-    run_kernel_col_reduction_cached(
-        exec,
-        [] GKO_KERNEL(auto i, auto j, auto x, auto y) {
-            return x(i, j) * y(i, j);
-        },
-        GKO_KERNEL_REDUCE_SUM(ValueType), result->get_values(), x->get_size(),
-        tmp, x, y);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_DOT_KERNEL);
-
-
-template <typename ValueType>
-void compute_conj_dot(std::shared_ptr<const DefaultExecutor> exec,
-                      const matrix::Dense<ValueType>* x,
-                      const matrix::Dense<ValueType>* y,
-                      matrix::Dense<ValueType>* result, array<char>& tmp)
-{
-    run_kernel_col_reduction_cached(
-        exec,
-        [] GKO_KERNEL(auto i, auto j, auto x, auto y) {
-            return conj(x(i, j)) * y(i, j);
-        },
-        GKO_KERNEL_REDUCE_SUM(ValueType), result->get_values(), x->get_size(),
-        tmp, x, y);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_CONJ_DOT_KERNEL);
-
-
-template <typename ValueType>
-void compute_norm2(std::shared_ptr<const DefaultExecutor> exec,
-                   const matrix::Dense<ValueType>* x,
-                   matrix::Dense<remove_complex<ValueType>>* result,
-                   array<char>& tmp)
-{
-    run_kernel_col_reduction_cached(
-        exec,
-        [] GKO_KERNEL(auto i, auto j, auto x) { return squared_norm(x(i, j)); },
-        [] GKO_KERNEL(auto a, auto b) { return a + b; },
-        [] GKO_KERNEL(auto a) { return sqrt(a); }, remove_complex<ValueType>{},
-        result->get_values(), x->get_size(), tmp, x);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_NORM2_KERNEL);
-
-template <typename ValueType>
-void compute_norm1(std::shared_ptr<const DefaultExecutor> exec,
-                   const matrix::Dense<ValueType>* x,
-                   matrix::Dense<remove_complex<ValueType>>* result,
-                   array<char>& tmp)
-{
-    run_kernel_col_reduction_cached(
-        exec, [] GKO_KERNEL(auto i, auto j, auto x) { return abs(x(i, j)); },
-        GKO_KERNEL_REDUCE_SUM(remove_complex<ValueType>), result->get_values(),
-        x->get_size(), tmp, x);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_NORM1_KERNEL);
-
-
-template <typename ValueType>
-void compute_max_nnz_per_row(std::shared_ptr<const DefaultExecutor> exec,
-                             const matrix::Dense<ValueType>* source,
-                             size_type& result)
-{
-    array<size_type> partial{exec, source->get_size()[0] + 1};
-    count_nonzeros_per_row(exec, source, partial.get_data());
-    run_kernel_reduction(
-        exec, [] GKO_KERNEL(auto i, auto partial) { return partial[i]; },
-        GKO_KERNEL_REDUCE_MAX(size_type),
-        partial.get_data() + source->get_size()[0], source->get_size()[0],
-        partial);
-    result = exec->copy_val_to_host(partial.get_const_data() +
-                                    source->get_size()[0]);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
-    GKO_DECLARE_DENSE_COMPUTE_MAX_NNZ_PER_ROW_KERNEL);
-
-
-template <typename ValueType>
-void compute_slice_sets(std::shared_ptr<const DefaultExecutor> exec,
-                        const matrix::Dense<ValueType>* source,
-                        size_type slice_size, size_type stride_factor,
-                        size_type* slice_sets, size_type* slice_lengths)
-{
-    const auto num_rows = source->get_size()[0];
-    array<size_type> row_nnz{exec, num_rows};
-    count_nonzeros_per_row(exec, source, row_nnz.get_data());
-    const auto num_slices =
-        static_cast<size_type>(ceildiv(num_rows, slice_size));
-    run_kernel_row_reduction(
-        exec,
-        [] GKO_KERNEL(auto slice, auto local_row, auto row_nnz, auto slice_size,
-                      auto stride_factor, auto num_rows) {
-            const auto row = slice * slice_size + local_row;
-            return row < num_rows ? static_cast<size_type>(
-                                        ceildiv(row_nnz[row], stride_factor) *
-                                        stride_factor)
-                                  : size_type{};
-        },
-        GKO_KERNEL_REDUCE_MAX(size_type), slice_lengths, 1,
-        gko::dim<2>{num_slices, slice_size}, row_nnz, slice_size, stride_factor,
-        num_rows);
-    exec->copy(num_slices, slice_lengths, slice_sets);
-    components::prefix_sum_nonnegative(exec, slice_sets, num_slices + 1);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
-    GKO_DECLARE_DENSE_COMPUTE_SLICE_SETS_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void count_nonzeros_per_row(std::shared_ptr<const DefaultExecutor> exec,
-                            const matrix::Dense<ValueType>* mtx,
-                            IndexType* result)
-{
-    run_kernel_row_reduction(
-        exec,
-        [] GKO_KERNEL(auto i, auto j, auto mtx) {
-            return is_nonzero(mtx(i, j)) ? 1 : 0;
-        },
-        GKO_KERNEL_REDUCE_SUM(IndexType), result, 1, mtx->get_size(), mtx);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_DENSE_COUNT_NONZEROS_PER_ROW_KERNEL);
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
-    GKO_DECLARE_DENSE_COUNT_NONZEROS_PER_ROW_KERNEL_SIZE_T);
-
-
-template <typename ValueType>
-void compute_squared_norm2(std::shared_ptr<const DefaultExecutor> exec,
-                           const matrix::Dense<ValueType>* x,
-                           matrix::Dense<remove_complex<ValueType>>* result,
-                           array<char>& tmp)
-{
-    run_kernel_col_reduction_cached(
-        exec,
-        [] GKO_KERNEL(auto i, auto j, auto x) { return squared_norm(x(i, j)); },
-        GKO_KERNEL_REDUCE_SUM(remove_complex<ValueType>), result->get_values(),
-        x->get_size(), tmp, x);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
-    GKO_DECLARE_DENSE_COMPUTE_SQUARED_NORM2_KERNEL);
-
-
-template <typename ValueType>
-void compute_sqrt(std::shared_ptr<const DefaultExecutor> exec,
-                  matrix::Dense<ValueType>* x)
-{
-    run_kernel(
-        exec,
-        [] GKO_KERNEL(auto row, auto col, auto x) {
-            x(row, col) = sqrt(x(row, col));
-        },
-        x->get_size(), x);
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_SQRT_KERNEL);
 
 
 template <typename ValueType, typename IndexType>
